:bug: Add missing deps import statement handler logic
